@@ -1,3 +1,3 @@
 module UserImpersonate
-  VERSION = "0.11.0"
+  VERSION = "0.12.0"
 end
