@@ -10,7 +10,7 @@ module TestUtils
   def rails_version_in_range?(minimum_version, maximum_version)
     minimum_version = Gem::Version.new(minimum_version) if minimum_version.is_a?(String)
     maximum_version = Gem::Version.new(maximum_version) if maximum_version.is_a?(String)
-    minimum_version <= rails_version && rails_version < maximum_version
+    minimum_version <= rails_version && rails_version <= maximum_version
   end
 
   def dummy_app_dir
@@ -37,6 +37,6 @@ module TestUtils
   end
 
   def rails_4?
-    rails_version_in_range?('4.0.0', '5.0.0')
+    rails_version_in_range?('4.0.0', '5.1.3')
   end
 end
